rename timer.h/cpp
