+++ source/core/util/Timer.cpp
@@ -1,4 +1,4 @@
-#include <core/util/timer.h>
+#include <core/util/Timer.h>
 #include <thread>
 #include <cmath>
