+++ source/producers/ESSStream/mo01_parser.cpp
@@ -1,6 +1,6 @@
 #include <producers/ESSStream/mo01_parser.h>
 
-#include <core/util/timer.h>
+#include <core/util/Timer.h>
 #include <core/util/logger.h>
 
 mo01_nmx::mo01_nmx()
