+++ tests/core/util/CMakeLists.txt
@@ -6,7 +6,7 @@ set(SOURCES
   ${dir}/h5json.cpp
   ${dir}/json_file.cpp
   ${dir}/lexical_extensions.cpp
-  ${dir}/timer.cpp
+  ${dir}/TimerTest.cpp
   ${dir}/string_extensions.cpp
   ${dir}/time_extensions.cpp
   ${dir}/UTF_extensions.cpp
